move stuff in more relevant places
